Moved MessageBubble to components directory
